requirements.txt
allure-pytest
allure-python-commons
apipkg
astor
atomicwrites
attrs
certifi
chardet
colorama
configparser
crayons
execnet
idna
iniconfig
openpyxl
packaging
pluggy
py
pyparsing
pytest
pytest-ast-transformer
pytest-forked
pytest-html
pytest-metadata
pytest-soft-assertions
pytest-xdist
requests
selenium
six
texttable
toml
urllib3
webdriver-manager
